Adds a offline delay to the FritzBoxTracker
Comes in handy when your devices go sleep and drop connectivity
temporarily. If set to > 0 the devices will be considered online
for the specified amount of intervals. If you set it to 2
and your interval is 1m they will be considered online for additional
two minutes before they are marked as not connected.