Removes engines except for AsyncEngine (#16)
I've decided to remove all engines except for the `AsyncEngine`.
Due to some tests this engine is amazingly stable, has great
performance and you do not need to think about synchronizing
parallel tasks stuff so much.

This decision was basically driven from a maintenance view of
perspective. In the future I like to add some infrastructural
code like an api to communicate with the engine. And I will
not be able to integrate necessary changes to all of the engines.